pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "egy-maintenance-scraper"
version = "0.1.0"
description = "Collect public listings for maintenance technicians in Egypt and export to CSV"
authors = [{name="Your Name"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx==0.27.2",
    "beautifulsoup4==4.12.3",
    "selectolax==0.3.21",
    "lxml==5.3.0",
    "tqdm==4.66.5",
    "typer==0.12.5",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "pandas==2.2.3",
    "phonenumberslite==8.13.50",
    "rapidfuzz==3.9.7",
    "tenacity==9.0.0",
]

[project.scripts]
egy-maint-scrape = "src.cli:app"
